Rename BaseObservable to TypeScript
