+++ src/observable/list/BaseObservableList.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservable} from "../BaseObservable.js";
+import {BaseObservable} from "../BaseObservable";
 
 export class BaseObservableList extends BaseObservable {
     emitReset() {
