+++ src/observable/map/BaseObservableMap.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservable} from "../BaseObservable.js";
+import {BaseObservable} from "../BaseObservable";
 
 export class BaseObservableMap extends BaseObservable {
     emitReset() {
